Added warning inside ProgressController
This warning will be displayed if the RequestProvider is null but the
user is still using the cancel function. Normally the user would except
that his calls are all canceled. If the RequestProvider is null, this
won't be the case. Therefore the developer will see a warning telling
him that he should consider using a RequestProvider implementation"