Renamed several classes + refactored ResponseService callbacks
The ResponseService callbacks will now include the request reference instead of the response. The response only makes sense for the onRequestError callback (due to the http error etc.). For the other cases it is more interesting to see which url, sql etc has been used. It also has the ability to be casted to the correct request class (which will probably explained via wiki).