Renamed all sql references to data + added new method to the cache adapter
The new method for the cache adapter is optional and can be used to define the cache key. By default the method is using the NetworkFutureRequest.fullUrl as a key.

This commit also reduces the amount of places that enforces certain rules. That should make the usage of this lib a bit more user friendly.