Added support for a return statement in DATA annotations
This return statement should give devs the option to do a data operation (like UPDATE) and return fresh data with the same future object. That should simplify code and align better with the recent changes of the required body type. This statement will only be used if not empty or the override flag is set to true.